test: conn module unit tests split
